Move indexing prompts to root
