+++ graphrag/index/graph/extractors/__init__.py
@@ -3,16 +3,13 @@
 
 """The Indexing Engine graph extractors package root."""
 
-from .claims import CLAIM_EXTRACTION_PROMPT, ClaimExtractor
+from .claims import ClaimExtractor
 from .community_reports import (
-    COMMUNITY_REPORT_PROMPT,
     CommunityReportsExtractor,
 )
 from .graph import GraphExtractionResult, GraphExtractor
 
 __all__ = [
-    "CLAIM_EXTRACTION_PROMPT",
-    "COMMUNITY_REPORT_PROMPT",
     "ClaimExtractor",
     "CommunityReportsExtractor",
     "GraphExtractionResult",
