+++ graphrag/index/graph/extractors/claims/__init__.py
@@ -4,6 +4,5 @@
 """The Indexing Engine graph extractors claims package root."""
 
 from .claim_extractor import ClaimExtractor
-from .prompts import CLAIM_EXTRACTION_PROMPT
 
-__all__ = ["CLAIM_EXTRACTION_PROMPT", "ClaimExtractor"]
+__all__ = ["ClaimExtractor"]
