+++ graphrag/index/graph/extractors/community_reports/__init__.py
@@ -8,7 +8,6 @@ import graphrag.index.graph.extractors.community_reports.schemas as schemas
 from .build_mixed_context import build_mixed_context
 from .community_reports_extractor import CommunityReportsExtractor
 from .prep_community_report_context import prep_community_report_context
-from .prompts import COMMUNITY_REPORT_PROMPT
 from .sort_context import sort_context
 from .utils import (
     filter_claims_to_nodes,
@@ -20,7 +19,6 @@ from .utils import (
 )
 
 __all__ = [
-    "COMMUNITY_REPORT_PROMPT",
     "CommunityReportsExtractor",
     "build_mixed_context",
     "filter_claims_to_nodes",
