+++ graphrag/index/graph/extractors/community_reports/community_reports_extractor.py
@@ -11,8 +11,7 @@ from typing import Any
 from graphrag.index.typing import ErrorHandlerFn
 from graphrag.index.utils import dict_has_keys_with_types
 from graphrag.llm import CompletionLLM
-
-from .prompts import COMMUNITY_REPORT_PROMPT
+from graphrag.prompts.index.community_report import COMMUNITY_REPORT_PROMPT
 
 log = logging.getLogger(__name__)
