+++ graphrag/index/graph/extractors/graph/__init__.py
@@ -8,11 +8,9 @@ from .graph_extractor import (
     GraphExtractionResult,
     GraphExtractor,
 )
-from .prompts import GRAPH_EXTRACTION_PROMPT
 
 __all__ = [
     "DEFAULT_ENTITY_TYPES",
-    "GRAPH_EXTRACTION_PROMPT",
     "GraphExtractionResult",
     "GraphExtractor",
 ]
