+++ graphrag/index/graph/extractors/summarize/__init__.py
@@ -7,6 +7,5 @@ from .description_summary_extractor import (
     SummarizationResult,
     SummarizeExtractor,
 )
-from .prompts import SUMMARIZE_PROMPT
 
-__all__ = ["SUMMARIZE_PROMPT", "SummarizationResult", "SummarizeExtractor"]
+__all__ = ["SummarizationResult", "SummarizeExtractor"]
