+++ graphrag/index/graph/extractors/summarize/description_summary_extractor.py
@@ -9,8 +9,7 @@ from dataclasses import dataclass
 from graphrag.index.typing import ErrorHandlerFn
 from graphrag.index.utils.tokens import num_tokens_from_string
 from graphrag.llm import CompletionLLM
-
-from .prompts import SUMMARIZE_PROMPT
+from graphrag.prompts.index.summarize_descriptions import SUMMARIZE_PROMPT
 
 # Max token size for input prompts
 DEFAULT_MAX_INPUT_TOKENS = 4_000
