+++ graphrag/prompts/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) 2024 Microsoft Corporation.
+# Licensed under the MIT License
+
+"""All prompts for the system."""
